+++ Roedor.cs
@@ -17,7 +17,7 @@ public class Roedor : Animal
         NoEspecificado
     }
 
-    public Roedor(string nombre, Cliente dueño, DateTime fecNac, int peso, Genero sexo, Variedad especie)
+    public Roedor(string nombre, Cliente dueño, DateTime fecNac, double peso, Genero sexo, Variedad especie)
         : base(nombre, dueño, fecNac, peso, sexo)
     {
         this.especie = especie;
